requirements.txt
attrs==19.3.0
more-itertools==8.0.2
numpy==1.18.0
packaging==19.2
Pint==0.9
pluggy==0.13.1
py==1.8.0
pybind11==2.4.3
pyfrc==2019.1.0
pynetconsole==2.0.2
pynetworktables==2019.0.2
pyparsing==2.4.5
pytest==5.3.2
robotpy-ctre==2019.3.2
robotpy-hal-base==2019.2.3
robotpy-hal-sim==2019.2.3
robotpy-installer==2019.2.0
robotpy-pathfinder==0.2.6
robotpy-rev==2019.0.2
robotpy-wpilib-utilities==2019.0.9
robotpy-navx==2019.0.1
six==1.13.0
wcwidth==0.1.7
wpilib==2019.2.3
wpilib.kinematics==0.1
